# backend/docs/swagger.yaml
definitions:
  github_com_lshc_contract-hub_backend_pkg_response.Body:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  internal_handler.LoginRequest:
    properties:
      password:
        description: 登录密码
        example: "123456"
        type: string
      username:
        description: 登录账号
        example: zhangsan
        type: string
    type: object
  internal_handler.RegisterRequest:
    properties:
      email:
        description: 邮箱
        example: zhangsan@example.com
        type: string
      nickname:
        description: 昵称/姓名
        example: 张三
        type: string
      password:
        description: 登录密码
        example: "123456"
        type: string
      phone:
        description: 手机号
        example: "13800138000"
        type: string
      username:
        description: 登录账号
        example: zhangsan
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 心智协同合同协作系统 V1 后端接口文档
  title: 心智协同合同协作系统 API
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: 校验用户名和密码，登录成功返回 JWT
      parameters:
      - description: 登录参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
        "401":
          description: 用户名或密码错误
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
      summary: 用户登录
      tags:
      - 认证
  /api/auth/me:
    get:
      consumes:
      - application/json
      description: 通过 JWT 获取当前登录用户信息
      produces:
      - application/json
      responses:
        "200":
          description: 当前用户信息
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
        "401":
          description: 未登录或 token 失效
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 获取当前用户
      tags:
      - 认证
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: 注册一个新的系统用户，用户名唯一，密码使用 bcrypt 加密保存
      parameters:
      - description: 注册参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 注册成功
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
        "409":
          description: 用户名已存在
          schema:
            $ref: '#/definitions/github_com_lshc_contract-hub_backend_pkg_response.Body'
      summary: 用户注册
      tags:
      - 认证
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"